pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "ats-resume-analyzer"
version = "0.2.1"
description = "ATS resume analyzer: PDF->text, parsing, TF-IDF/semantic scoring, LLM suggestions, prompt generation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "streamlit>=1.37",
  "typer>=0.12",
  "pymupdf>=1.24",
  "Pillow>=10.0",
  "pytesseract>=0.3",
  "pandas>=2.2",
  "scikit-learn>=1.4",
  "rapidfuzz>=3.9",
  "pydantic>=2.6",
  "pydantic-settings>=2.3",
  "python-dotenv>=1.0",
  "litellm>=1.43",
  "instructor>=1.5",
  "python-docx>=0.8.11",
  "sentence-transformers>=2.6",
  "transformers>=4.44",
  "huggingface_hub>=0.24",
  # YAML + safe parsing for config
  "pyyaml>=6.0.1"
]

[dependency-groups]
dev = [
  "pytest>=8.2",
  "pre-commit>=4.0",
  "ruff>=0.6",
  "black>=24.8",
]
# Optional heavier NLP utilities you can sync when needed
nlp = [
  "spacy>=3.7",
  "spacy-transformers>=1.3",
]

[tool.hatch.build.targets.wheel]
# src layout – package lives under src/ats
packages = ["src/ats"]

[tool.hatch.build.targets.sdist]
include = [
  "src/ats",
  "README.md",
  "config/**",
]

[tool.ruff]
line-length = 100
